+++ engine/code/bspc/l_bsp_sin.c
@@ -832,7 +832,7 @@ void	Sin_LoadBSPFile(char *filename, int offset, int length)
 	LoadFile (filename, (void **)&header, offset, length);
 
 // swap the header
-	for (i=0 ; (size_t)i< sizeof(sin_dheader_t)/4 ; i++)
+	for (i=0 ; i< sizeof(sin_dheader_t)/4 ; i++)
 		((int *)header)[i] = LittleLong ( ((int *)header)[i]);
 
 	if (header->ident != SIN_BSPHEADER && header->ident != SINGAME_BSPHEADER)
@@ -914,7 +914,7 @@ void	Sin_LoadBSPFileTexinfo (char *filename)
 	fread (header, sizeof(sin_dheader_t), 1, f);
 
 // swap the header
-	for (i=0 ; (size_t)i< sizeof(sin_dheader_t)/4 ; i++)
+	for (i=0 ; i< sizeof(sin_dheader_t)/4 ; i++)
 		((int *)header)[i] = LittleLong ( ((int *)header)[i]);
 
 	if (header->ident != SIN_BSPHEADER && header->ident != SINGAME_BSPHEADER)
